requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.21
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
python-dotenv==1.1.0
typing_extensions==4.13.2
